feat: wip crop probability admin page
