+++ apps/picsa-apps/dashboard/src/app/components/data-import/data-import.component.scss
@@ -1,0 +1,3 @@
+:host {
+  display: block;
+}
